+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Non-bound properties inside ViewModels can locate it's ValidationContext, and hence can now be used with validators
 - SortedFilteredList will not overwrite the backing list when column sorting is enabled in TableView (setAll disabled) (https://github.com/edvin/tornadofx/issues/344)
 - RowExpanders containing nested TableViews no longer inherits white label text when owning row is selected
+- Calling `cellFormat` on a TableCell that already has a formatter will now add the new formatter as a decorator instead of overwriting the old
 
 ### Changes